+++ src/compas_fab/rhino/__init__.py
@@ -13,14 +13,28 @@ datastructures and models, in a way that maintains the data separated from the
 specific CAD interfaces, while providing a way to leverage native performance
 of the CAD environment.
 
-This package only provides a deprecated alias to the
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    ReachabilityMapArtist
+
+Additionally, this package provides deprecated aliases to the
 :class:`~compas_rhino.artists.RobotModelArtist` implemented in COMPAS.
 
 """
 import compas
 
 if compas.RHINO:
-    from .artists import RobotArtist
-    from .artists import RobotModelArtist
+    from .artists import (
+        ReachabilityMapArtist,
+        RobotArtist,
+        RobotModelArtist,
+    )
 
-    __all__ = ['RobotArtist', 'RobotModelArtist']
+    __all__ = [
+        'ReachabilityMapArtist',
+
+        'RobotArtist',              # deprecated
+        'RobotModelArtist',         # deprecated
+    ]