+++ src/compas_fab/robots/reachability_map/__init__.py
@@ -1,0 +1,9 @@
+from .reachability_map import ReachabilityMap
+from .vector_generators import DeviationVectorsGenerator
+from .vector_generators import OrthonormalVectorsFromAxisGenerator
+
+__all__ = [
+    'ReachabilityMap',
+    'DeviationVectorsGenerator',
+    'OrthonormalVectorsFromAxisGenerator'
+]